Fix pointer-sized share prefix count marshalling
Agent-Logs-Url: https://github.com/TopperDEL/uplink.NET.New/sessions/0bbfc11a-42c2-4150-8f6b-b2d255826751

+++ src/uplink.NET/Native/UplinkInterop.cs
@@ -279,7 +279,7 @@ internal static unsafe partial class UplinkInterop
         nint access,
         UplinkPermission permission,
         UplinkSharePrefix* prefixes,
-        int prefixes_count);
+        nint prefixes_count);
 
     [LibraryImport(LibName)]
     internal static partial void uplink_free_string_result(UplinkStringResult result);
